добавил фрагменты (без Табов)
