+++ app/src/main/res/values-ru/strings.xml
@@ -3,5 +3,7 @@
     <string name="edittext_name_hint">Название</string>
     <string name="edittext_price_hint">Цена</string>
     <string name="add_button_title">Добавить</string>
+    <string name="add">Добавить</string>
+    <string name="plusAddButton">+</string>
 
 </resources>
