+++ app/src/main/res/values/strings.xml
@@ -3,5 +3,7 @@
     <string name="edittext_name_hint">Title</string>
     <string name="edittext_price_hint">Price</string>
     <string name="add_button_title">Add</string>
+    <string name="add">Add</string>
+    <string name="plusAddButton">+</string>
 
 </resources>
